Code organisation and refactoring
Code organised into two python modules. Obsolete code removed and remaining code refactored.
