+++ client/templates/terms_and_conditions.html
