+++ run-client.py
@@ -1,0 +1,4 @@
+from client import app
+
+if __name__ == "__main__":
+    app.run(port=5000, debug=True)
