+++ run-server.py
@@ -1,0 +1,4 @@
+from server import app
+
+if __name__ == "__main__":
+    app.run(port=5001, debug=True)
