+++ server/__init__.py
@@ -1,0 +1,7 @@
+from flask import Flask
+from flask_restful import Api
+
+app = Flask(__name__)
+api = Api(app)
+
+from server import routes
